+++ src/packages/piral-core/src/state/actions/search.ts
@@ -1,6 +1,7 @@
 import { ReactChild } from 'react';
-import { swap, Atom } from '@dbeining/react-atom';
-import { GlobalState } from '../../types';
+import { wrapElement } from 'react-arbiter';
+import { swap, Atom, deref } from '@dbeining/react-atom';
+import { GlobalState, Disposable, SearchOptions } from '../../types';
 import { appendItems, prependItems } from '../../utils';
 
 export function setSearchInput(input: string) {
@@ -13,12 +14,60 @@ export function setSearchInput(input: string) {
   }));
 }
 
-export function resetSearchResults(loading: boolean) {
+export function triggerSearch(query?: string, immediate = false): Disposable {
+  const state = deref(this as Atom<GlobalState>);
+  const providers = state.components.searchProviders;
+  const { input, loading } = state.search;
+
+  if (query === undefined) {
+    query = input;
+  }
+
+  if (input !== query || !loading) {
+    const allProviders = Object.keys(providers);
+    const providerKeys = allProviders.filter(m => !providers[m].onlyImmediate || immediate);
+    const load = !!query && providerKeys.length > 0;
+    resetSearchResults.call(this, query, load);
+
+    if (load) {
+      let searchCount = providerKeys.length;
+      let active = true;
+      const opts: SearchOptions = {
+        query,
+        immediate,
+      };
+
+      providerKeys.forEach(key =>
+        providers[key].search(opts).then(
+          results => {
+            active && appendSearchResults(results.map(m => wrapElement(m)), --searchCount === 0);
+          },
+          ex => {
+            console.warn(ex);
+            active && --searchCount === 0 && appendSearchResults([], true);
+          },
+        ),
+      );
+
+      return () => {
+        active = false;
+        providerKeys.forEach(key => providers[key].cancel());
+      };
+    } else if (!query) {
+      allProviders.forEach(key => providers[key].clear());
+    }
+  }
+
+  return () => {};
+}
+
+export function resetSearchResults(input: string, loading: boolean) {
   swap(this as Atom<GlobalState>, state => ({
     ...state,
     search: {
       ...state.search,
       loading,
+      input,
       results: [],
     },
   }));